requirements.txt
boto3==1.22.2
botocore==1.25.2
click==8.1.2
Flask==2.1.1
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
jmespath==1.0.0
MarkupSafe==2.1.1
protobuf==3.20.1
python-dateutil==2.8.2
python-dotenv==0.20.0
s3transfer==0.5.2
six==1.16.0
urllib3==1.26.9
Werkzeug==2.1.1
zipp==3.8.0
